Re-work interfaces to abstract Segment v Index.
